Criteria. Extract inmemory evaluator into a separate sub-module: inmemory
