+++ criteria/inmemory/README.md
@@ -1,0 +1,3 @@
+In memory adapter that uses reflection to evaluate an expression.
+
+Evaluator can be converted to a Predicate and applied to generic Iterable(s).
